Re-wrote the venn diagrams to use tikz
Removed the public domain wikimedia Venn diagrams in favor of tikz generated
images. The entire reason was that I wanted there to be labels and I wanted the
color to be gray. Oh man, this is a big display of working 30 minutes to solve a
problem that I could do have done with MS Paint... :-P