Tiny matrix change and added solving recurrences
The tiny matrix change makes a homework question a little more interesting -
we're finding the transitive closure of the relation (without saying that's what
we're doing) and the problem as previously written was a little more boring that
I wanted.

I added the section on solving recurrences with Tower of Hanoi, Catalan, a stamp
machine, and another. Next we'll be doing linear homogenous relations.