A first, crummy draft of an induction computer example
On the one hand, I like the spirit of it, but it's imprecise to the
point of being wrong.
And our students don't do algorithms until like senior year, so they
won't have seen QuickSort... Crumb. I need another computer algorithm
that they may be more familiar with.

Committing this to include it as a "we're working on it" idea.